requirements.txt
attrs==21.4.0
Automat==20.2.0
certifi==2021.10.8
chardet==4.0.0
click==7.1.2
ConfigArgParse==0.15.1
constantly==15.1.0
Flask==1.1.4
gevent==1.5a2
geventhttpclient-wheels==1.3.1.dev2
greenlet==1.1.2
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
itsdangerous==1.1.0
Jinja2==2.11.3
locustio==0.13.5
MarkupSafe==1.1.1
msgpack-python==0.5.6
pkg-resources==0.0.0
PyHamcrest==2.0.3
pyzmq==20.0.0
requests==2.25.1
six==1.16.0
Twisted==20.3.0
urllib3==1.26.9
Werkzeug==1.0.1
zope.interface==5.4.0
